added carosel HTML for landing page
