--- client/src/components/Landing.jsx
@@ -1,32 +1,0 @@
-import React from "react";
-import axios from "axios";
-import { useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
-
-const Landing = (props) => {
-  let history = useHistory();
-
-  let [registeredUser, setRegisteredUSer] = useState({});
-
-  useEffect(() => {
-    axios
-      .get("http://localhost:8000/api/user/getone", { withCredentials: true })
-      .then((res) => {
-        if (res.data) {
-          setRegisteredUSer(res.data);
-        }
-      })
-      .catch((err) => {
-        history.push("/");
-        console.log("ERR WHEN GETTING LOGGED IN USER", err);
-      });
-  }, []);
-
-  return (
-    <div>
-      <h1>INSERT CAROSEL HERE ------- FROM LANDING</h1>
-    </div>
-  );
-};
-
-export default Landing;
