+++ client/src/components/login-reg-components/Registration.jsx
@@ -2,7 +2,8 @@ import React, { useState } from "react";
 import axios from "axios";
 import { useHistory } from "react-router-dom";
 import Button from "@mui/material/Button";
-import { HeroImg } from "../HeroImg";
+import { HeroImg } from "./HeroImg";
+
 import style from "./login-reg-component-styles.css";
 
 const User = (props) => {
